Add files via upload
In this project, I use html, css and javascript to create a web app tic-tac-toe game that allow the user to play the game with a computer. The game checks the event of 3 inline horizontal, vertical and or diagonal matches for a win. The message declaring the winner at the end of the game is displayed upon game completion. The user can either then quit the game or click restart to restart the game.